coral-server/agents/reception/pyproject.toml
[project]
name = "reception"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.6.0",
    "groq>=0.31.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "langchain-mcp-adapters>=0.1.10",
    "langgraph>=0.6.7",
    "livekit-agents>=1.2.11",
    "livekit-plugins-cartesia>=1.2.11",
    "livekit-plugins-deepgram>=1.2.11",
    "livekit-plugins-groq>=1.2.11",
    "livekit-plugins-openai>=1.2.11",
    "livekit-plugins-silero>=1.2.11",
    "python-dotenv>=1.1.1",
    "uv>=0.8.18",
]
